Hooray, fully functional matrix

+++ transpose.go
@@ -7,15 +7,15 @@ type Transposed struct {
 // Verify Matric interface implementation
 var _ Matrix = &Transposed{}
 
-func (t *Transposed) Set(i, j uint32) error {
+func (t *Transposed) Set(i, j uint64) error {
 	return t.Matrix.Set(j, i)
 }
 
-func (t *Transposed) SetBit(i, j, k uint32) error {
+func (t *Transposed) SetBit(i, j, k uint64) error {
 	return t.Matrix.SetBit(j, i, k)
 }
 
-func (t *Transposed) Get(i, j uint32) (uint64, error) {
+func (t *Transposed) Get(i, j uint64) (uint64, error) {
 	return t.Matrix.Get(j, i)
 }
